Protocol version 2
Remove KademliaId message (is now inlined). This saves a few bytes for
messages using ids. In particular, User objects now take less space per
file.

Protocol version is now indicated in the first byte, not in the
UdpMessage or ServerHandshake.

This is an incompatible change and invalidates any v1 User objects.